Volume Surrogate/requirements.txt
# Field Surrogate - Python Dependencies
# =====================================
# Install all dependencies with: pip install -r requirements.txt

# Core Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Ansys Fluent Integration
ansys-fluent-core>=0.18.0

# Additional dependencies that may be required by ansys-fluent-core
grpcio>=1.50.0
protobuf>=4.21.0

# Optional but recommended
pandas>=2.0.0  # For data manipulation if needed
